Problem: Waiting for DDP can be slow (#372)
Solution: Implement Fast Render on all routes. Fix failing tests.
