+++ .meteor/packages
@@ -42,3 +42,4 @@ littledata:synced-cron
 pauldowman:dotenv
 socialize:user-presence
 easy:search
+staringatlights:fast-render@3.0.3
