+++ .meteor/versions
@@ -22,6 +22,7 @@ caching-html-compiler@1.1.2
 callback-hook@1.1.0
 cfs:graphicsmagick@0.0.18
 check@1.3.1
+chuangbo:cookie@1.1.0
 coffeescript@1.0.17
 ddp@1.4.0
 ddp-client@2.3.2
@@ -64,6 +65,8 @@ matb33:collection-hooks@0.7.15
 mdg:validated-method@1.1.0
 meteor@1.9.0
 meteor-base@1.4.0
+meteorhacks:meteorx@1.4.1
+meteorhacks:picker@1.0.3
 meteortesting:browser-tests@1.0.0
 meteortesting:mocha@1.0.0
 minifier-css@1.3.1
@@ -101,6 +104,7 @@ reactive-var@1.0.11
 reload@1.2.0
 retry@1.1.0
 routepolicy@1.0.13
+server-render@0.3.1
 service-configuration@1.0.11
 session@1.1.7
 sha@1.0.9
@@ -113,6 +117,8 @@ spacebars-compiler@1.1.3
 srp@1.0.10
 standard-minifier-css@1.4.1
 standard-minifier-js@2.3.4
+staringatlights:fast-render@3.0.4
+staringatlights:inject-data@2.2.0
 templating@1.3.2
 templating-compiler@1.3.3
 templating-runtime@1.3.2
