+++ imports/startup/both/index.js
@@ -1,2 +1,4 @@
 // Import modules used by both client and server through a single index entry point
 // e.g. useraccounts configuration file.
+
+import './routes'
